+++ src/client/__tests__/App.test.js
@@ -4,10 +4,6 @@ import {Provider} from 'react-redux';
 import createStore from '../redux/Store';
 import renderer from 'react-test-renderer';
 
-// jest.mock('../../data/exportTaxonomy.json', () => {
-//   return require('./__mocks__/mockedTaxonomy.json');
-// });
-
 jest.mock('../components/base/Belt/Slider.component', () => 'slider');
 jest.mock('../components/base/Belt/CreatorBelt.component', () => 'CreatorBelt');
 jest.mock('../components/base/Belt/TagsBelt.component', () => 'TagsBelt');
