+++ src/client/components/base/Belt/TagsBelt.component.js
@@ -3,16 +3,14 @@ import {get} from 'lodash';
 import Textarea from 'react-textarea-autosize';
 import {isMobileOnly} from 'react-device-detect';
 import Pin from '../Pin';
-import withIsVisible from '../Scroll/withIsVisible.hoc';
-import withChildBelt from './withChildBelt.hoc';
-import withStoreBelt from './withStoreBelt.hoc';
-import withTagsToPids from '../Recommender/withTagsToPids.hoc';
-import withUser from '../User/withUser.hoc';
+import {withIsVisible, withScrollToComponent} from '../../hoc/Scroll';
+import {withChildBelt, withStoreBelt} from '../../hoc/Belt';
+import {withTagsToPids} from '../../hoc/Recommender';
+import {withUser} from '../../hoc/User';
 import WorkSlider from './WorkSlider.component';
 import Title from '../Title';
 import Link from '../../general/Link.component';
 import Term from '../Term';
-import withScrollToComponent from '../Scroll/withScrollToComponent.hoc';
 import {filtersMapAll} from '../../../redux/filter.reducer';
 import ContextMenu, {ContextMenuAction} from '../ContextMenu';
 import Text from '../Text';
