+++ src/client/components/hoc/AdressBar/index.js
@@ -1,0 +1,7 @@
+import withTagsFromUrl, {
+  tagsFromURL,
+  addTag,
+  removeTag
+} from './withTagsFromUrl.hoc';
+
+export {withTagsFromUrl, tagsFromURL, addTag, removeTag};
