+++ src/client/components/hoc/Belt/index.js
@@ -1,0 +1,4 @@
+import withChildBelt from './withChildBelt.hoc';
+import withStoreBelt from './withStoreBelt.hoc';
+
+export {withChildBelt, withStoreBelt};
