+++ src/client/components/hoc/Belt/withStoreBelt.hoc.js
