+++ src/client/components/hoc/List/index.js
@@ -1,0 +1,1 @@
+export {withListCreator, withList} from './withList.hoc';
