+++ src/client/components/hoc/Recommender/index.js
@@ -1,0 +1,11 @@
+import withInteractionsToPids from './withInteractionsToPids.hoc';
+import withPidsToPids from './withPidsToPids.hoc';
+import withQueryToPids from './withQueryToPids.hoc';
+import withTagsToPids from './withTagsToPids.hoc';
+
+export {
+  withInteractionsToPids,
+  withPidsToPids,
+  withQueryToPids,
+  withTagsToPids
+};
