+++ src/client/components/hoc/Scroll/index.js
@@ -1,0 +1,4 @@
+import withIsVisible from './withIsVisible.hoc';
+import withScrollToComponent from './withScrollToComponent.hoc';
+
+export {withIsVisible, withScrollToComponent};
