+++ src/client/components/hoc/User/index.js
@@ -1,0 +1,3 @@
+import withUser from './withUser.hoc';
+
+export {withUser};
