+++ src/client/components/hoc/User/withUser.hoc.js
