+++ src/client/components/hoc/Work/index.js
@@ -1,0 +1,3 @@
+import withWork from './withWork.hoc';
+
+export {withWork};
