+++ src/client/components/hoc/Work/withWork.hoc.js
