+++ src/client/components/list/templates/bookcase/BookcaseList.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {connect} from 'react-redux';
 import scrollToComponent from 'react-scroll-to-component';
-import {withList} from '../../../base/List/withList.hoc';
+import {withList} from '../../../hoc/List';
 import AddToList from '../../addtolist/AddToList.container';
 import ListElement from '../../element/ListElement';
 import ListInfo from './ListInfo';
