+++ src/client/components/list/templates/simple/SimpleList.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {connect} from 'react-redux';
 import scrollToComponent from 'react-scroll-to-component';
-import {withList} from '../../../base/List/withList.hoc';
+import {withList} from '../../../hoc/List';
 import T from '../../../base/T';
 import Icon from '../../../base/Icon';
 import Text from '../../../base/Text';
